+++ tests/stubs/functions.php
@@ -18,13 +18,6 @@ if (!function_exists('acf_add_local_field_group')) {
     }
 }
 
-if (!function_exists('acf_add_options_page')) {
-    function acf_add_options_page()
-    {
-        //
-    }
-}
-
 if (!function_exists('get_field')) {
     function get_field($value, $post = null)
     {